# app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eth-tools API",
    "description": "Self-maintaining runtime for ERC-8004 trustless agents.",
    "license": {
      "name": "MIT"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "https://eth-tools.dev",
      "description": "Production"
    },
    {
      "url": "https://preview.eth-tools.dev",
      "description": "Preview"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    }
  ],
  "paths": {
...
    "/api/v1/invoke": {
      "post": {
        "tags": [
          "invoke"
        ],
        "operationId": "invoke_execute",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvokeExecute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Submitted (broadcast=true) or signed-only (broadcast=false)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "402": {
            "description": "x402 payment required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "x402 + wallet rails not yet wired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoke/prepare": {
      "post": {
        "tags": [
          "invoke"
        ],
        "operationId": "invoke_prepare",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvokePrepare"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Calldata + tx envelope, ready to sign",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvokePrepareResponse"
                }
              }
            }
          },
          "400": {
            "description": "bad input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "missing bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "chain not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Top-level error envelope returned by every non-2xx response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayload"
          }
        }
      },
      "ApiErrorPayload": {
        "type": "object",
        "required": [
          "code",
          "policy_version",
          "evaluator"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "AGENT_NOT_FOUND"
          },
          "evaluator": {
            "type": "string",
            "example": "api.lookup"
          },
          "override_hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "policy_version": {
            "type": "string",
            "example": "v1"
          }
        }
      },
...
      "InvokeExecute": {
        "type": "object",
        "required": [
          "chain",
          "agent_id",
          "to",
          "selector"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "args_hex": {
            "type": "string"
          },
          "broadcast": {
            "type": "boolean",
            "description": "Whether to broadcast immediately. When `false`, the wallet rails return\nthe signed tx in `signed_tx_hex` so the caller can broadcast itself."
          },
          "chain": {
            "type": "string"
          },
          "selector": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "value_wei": {
            "type": "string"
          }
        }
      },
      "InvokePrepare": {
        "type": "object",
        "required": [
          "chain",
          "agent_id",
          "to",
          "selector"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "args_hex": {
            "type": "string",
            "description": "Hex-encoded ABI args (concatenated, no `0x` prefix). Empty = no args."
          },
          "chain": {
            "type": "string"
          },
          "selector": {
            "type": "string",
            "description": "4-byte function selector (`0x12345678`)."
          },
          "to": {
            "type": "string",
            "description": "0x-prefixed contract address to call. The wallet rails restrict this\nto the three ERC-8004 registries; preview mode is permissive but warns."
          },
          "value_wei": {
            "type": "string"
          }
        }
      },
      "InvokePrepareResponse": {
        "type": "object",
        "required": [
          "calldata",
          "to",
          "value_wei",
          "chain_id",
          "agent_id"
        ],
        "properties": {
          "agent_id": {
            "type": "string",
            "description": "Echoed back so the caller can audit."
          },
          "calldata": {
            "type": "string",
            "description": "Fully-encoded calldata `0x` + selector + args. Hand this to the\nsigning rails (or a user's external wallet) — this endpoint never\nsigns anything."
          },
          "chain_id": {
            "type": "integer",
            "format": "int64"
          },
          "to": {
            "type": "string"
          },
          "value_wei": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}